requirements.txt
fastai==2.0.15
torch==1.6.0
scipy==1.5.2
pandas==1.0.1
spacy==2.3.1
nltk==3.5
prettytable==1.0.0
numpy==1.18.5
networkx==2.5
tqdm==4.49.0
scikit_learn==0.23.2
gensim>=4.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
PyYAML>=6.0
